[CLEAN] [MIGRATE] membership: cleaning of the model + migrate to the new API.
Future commits will aim to improve some features of the membership module.
However cleaning it before adding or changing features seems a good idea
due to the code quality not being very impressive currently.
