+++ addons/membership/models/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+import membership
+import product
+import account_invoice
+import res_partner
