+++ addons/membership/report/__init__.py
@@ -19,5 +19,5 @@
 #
 ##############################################################################
 
-import report_membership
+# import report_membership
 # vim:expandtab:smartindent:tabstop=4:softtabstop=4:shiftwidth=4:
